+++ src/com/company/dividedBy.java
@@ -3,7 +3,7 @@ package com.company;
 import java.util.Scanner;
 
 public class dividedBy {
-    public int divided(int x, int y) {
-        return x / y;
+    public float divided(float x,float  y) {
+        return (x / y);
     }
 }
