+++ src/com/company/multiply.java
@@ -1,7 +1,7 @@
 package com.company;
 
 public class multiply {
-    public int multiplication (int x, int y) {
+    public float multiplication (float x, float y) {
         return x * y;
     }
 }
